Moved to generic artifact dust type
